OASP4J Set the GenerationType to SEQUENCE in the class AbstractPersistenceEntity (jpa module) and including the Entity code in the class ApplicationPersistenceEntity of the sample application
